requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9
python-decouple>=3.8
requests>=2.31
geopy>=2.4
drf-spectacular>=0.27
gunicorn>=21.2
